+++ scripts/populate_tox/config.py
@@ -47,6 +47,12 @@ TEST_SUITE_CONFIG = {
         "package": "apache-beam",
         "python": ">=3.7",
     },
+    "boto3": {
+        "package": "boto3",
+        "deps": {
+            "py3.7,py3.8": ["urllib3<2.0.0"],
+        },
+    },
     "bottle": {
         "package": "bottle",
         "deps": {
@@ -56,8 +62,14 @@ TEST_SUITE_CONFIG = {
     "celery": {
         "package": "celery",
         "deps": {
-            "*": ["newrelic", "redis"],
+            "*": ["newrelic<10.17.0", "redis"],
             "py3.7": ["importlib-metadata<5.0"],
+        },
+    },
+    "chalice": {
+        "package": "chalice",
+        "deps": {
+            "*": ["pytest-chalice"],
         },
     },
     "clickhouse_driver": {
@@ -143,6 +155,9 @@ TEST_SUITE_CONFIG = {
     },
     "huggingface_hub": {
         "package": "huggingface_hub",
+        "deps": {
+            "*": ["responses"],
+        },
     },
     "langchain-base": {
         "package": "langchain",
@@ -218,6 +233,20 @@ TEST_SUITE_CONFIG = {
         "package": "pyramid",
         "deps": {
             "*": ["werkzeug<2.1.0"],
+        },
+    },
+    "quart": {
+        "package": "quart",
+        "deps": {
+            "*": ["quart-auth", "pytest-asyncio", "Werkzeug"],
+            ">=0.19": ["quart-flask-patch"],
+            "<0.19": [
+                "blinker<1.6",
+                "jinja2<3.1.0",
+                "Werkzeug<2.3.0",
+                "hypercorn<0.15.0",
+            ],
+            "py3.8": ["taskgroup==0.0.0a4"],
         },
     },
     "redis_py_cluster_legacy": {